Добавлен выбор y/n при большом количестве вариантов подстановок
